+++ app/src/main/java/in/khatri/rahul/amiiboapp/java/retrofit/HomeRetrofitActivity.java
@@ -17,17 +17,12 @@ import android.widget.EditText;
 import android.widget.LinearLayout;
 import android.widget.Toast;
 
-import com.google.gson.JsonElement;
-
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
 import java.util.ArrayList;
 
 import in.khatri.rahul.amiiboapp.R;
-import in.khatri.rahul.amiiboapp.java.adapter.GameAdapter;
-import in.khatri.rahul.amiiboapp.java.fastNetworking.model.GameModel;
+import in.khatri.rahul.amiiboapp.java.retrofit.adapter.GameRetrofitAdapter;
+import in.khatri.rahul.amiiboapp.java.retrofit.model.GameDataList;
+import in.khatri.rahul.amiiboapp.java.retrofit.model.GameRetrofitModel;
 import in.khatri.rahul.amiiboapp.java.retrofit.utils.ApiInterface;
 import in.khatri.rahul.amiiboapp.java.retrofit.utils.ClientApi;
 import in.khatri.rahul.amiiboapp.java.utils.dialog.SpotsDialog;
@@ -41,13 +36,10 @@ public class HomeRetrofitActivity extends AppCompatActivity {
     RecyclerView recyclerView;
     LinearLayout llNoData;
     SpotsDialog progressDialog;
-    GameModel gameModel;
-    GameAdapter gameAdapter = null;
-    ArrayList<GameModel> arrayListGame = new ArrayList<>();
+    GameRetrofitAdapter gameAdapter = null;
+    ArrayList<GameRetrofitModel> arrayListGame = new ArrayList<>();
     private ApiInterface apiInterface;
     boolean doubleBackToExitPressedOnce = false;
-    private final String BASE_URL = "https://www.amiiboapi.com/api/";
-
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -66,11 +58,9 @@ public class HomeRetrofitActivity extends AppCompatActivity {
         if (!isOnline()) {
             checkNetwork();
         }
-        //  Retrofit retrofit = new Retrofit.Builder().addConverterFactory(GsonConverterFactory.create()).baseUrl(BASE_URL).build();
-        //  apiInterface = retrofit.create(ApiInterface.class);
         apiInterface= ClientApi.getClient().create(ApiInterface.class);
         gameData();
-        gameAdapter= new GameAdapter(this,arrayListGame);
+        gameAdapter= new GameRetrofitAdapter(this, arrayListGame);
         etSearch.addTextChangedListener(new TextWatcher() {
             @Override
             public void beforeTextChanged(CharSequence s, int start, int count, int after) {
@@ -81,7 +71,7 @@ public class HomeRetrofitActivity extends AppCompatActivity {
             public void onTextChanged(CharSequence s, int start, int before, int count) {
                 Log.e("et","log");
                 if (!s.equals("")) {
-                    ArrayList<GameModel> searchList = new ArrayList();
+                    ArrayList<GameRetrofitModel> searchList = new ArrayList();
                     for (int i = 0; i < arrayListGame.size(); i++) {
                         if (arrayListGame.get(i).getName().toLowerCase().contains(s) || arrayListGame.get(i).getGameSeries().toLowerCase().contains(s)) {
                             searchList.add(arrayListGame.get(i));
@@ -89,14 +79,15 @@ public class HomeRetrofitActivity extends AppCompatActivity {
                     }
                     try {
                         gameAdapter.notifyDataSetChanged();
-                        recyclerView.setAdapter(new GameAdapter(HomeRetrofitActivity.this, searchList));
+                        recyclerView.setAdapter(new GameRetrofitAdapter(HomeRetrofitActivity.this, searchList));
                     } catch (Exception e) {
                         e.printStackTrace();
                     }
 
                 } else {
                     gameAdapter.notifyDataSetChanged();
-                    recyclerView.setAdapter(new GameAdapter(HomeRetrofitActivity.this, arrayListGame));
+                    recyclerView.setAdapter(new GameRetrofitAdapter(HomeRetrofitActivity.this, arrayListGame));
+
                 }
             }
 
@@ -109,89 +100,29 @@ public class HomeRetrofitActivity extends AppCompatActivity {
 
     private void gameData() {
         arrayListGame.clear();
-        Call<JsonElement> call = apiInterface.getGameData();
-        call.enqueue(new Callback<JsonElement>() {
+        Call<GameDataList> call = apiInterface.getGameData();
+        call.enqueue(new Callback<GameDataList>() {
             @Override
-            public void onResponse(Call<JsonElement> call, Response<JsonElement> response) {
+            public void onResponse(Call<GameDataList> call, Response<GameDataList> response) {
                 Log.e("res body", response.body() + "");
-                String stResponse = String.valueOf(response.body());
-                Log.e("StRes", stResponse);
-              /*  gameAdapter= new GameRetrofitAdapter(HomeRetrofitActivity.this, response.body());
-                Type listType= new TypeToken<List<GameRetrofitModel>>() {}.getType();
-
-                 arrayListGame=new Gson().fromJson(stResponse,listType);
-                Log.e("arr ", arrayListGame+"" );
-                Log.e("arr size", arrayListGame.size()+"" );*/
-                getGameData(stResponse);
+                arrayListGame = response.body().getDataList();
+//                String model = "ammio chrac "+arrayList.toString()+" series ";// +gameRetrofitModel.getGameSeries();//+" au "+response.body().getRelease().getAu();
+//                Log.e("response retrofit", model);
+                gameAdapter = new GameRetrofitAdapter(HomeRetrofitActivity.this, arrayListGame);
+                recyclerView.setAdapter(gameAdapter);
+                recyclerView.setVisibility(View.VISIBLE);
+                llNoData.setVisibility(View.GONE);
                 progressDialog.dismiss();
             }
 
             @Override
-            public void onFailure(Call<JsonElement> call, Throwable t) {
+            public void onFailure(Call<GameDataList> call, Throwable t) {
                 Log.e("HomeRetofitAct", "FailureOccur:" + t);
+                llNoData.setVisibility(View.VISIBLE);
+                recyclerView.setVisibility(View.GONE);
                 progressDialog.dismiss();
             }
         });
-
-    }
-
-    private void getGameData(String response) {
-        try {
-            JSONObject jsonObject = new JSONObject(response);
-            final JSONArray jsonArray = jsonObject.getJSONArray("amiibo");
-            if (this != null) {
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        if (jsonArray.length() > 0) {
-                            for (int i = 0; i < jsonArray.length(); i++) {
-                                try {
-                                    JSONObject jsonObject1 = jsonArray.getJSONObject(i);
-                                    JSONObject jsonObjectRelease = jsonObject1.getJSONObject("release");
-                                    gameModel = new GameModel(jsonObject1.getString("amiiboSeries"), jsonObject1.getString("character"), jsonObject1.getString("gameSeries"), jsonObject1.getString("head"), jsonObject1.getString("image"), jsonObject1.getString("name"), jsonObject1.getString("tail"), jsonObject1.getString("type"),jsonObjectRelease.getString("au"), jsonObjectRelease.getString("eu"), jsonObjectRelease.getString("jp"), jsonObjectRelease.getString("na"));
-                                    /*gameModel.setAmiiboSeries(jsonObject1.getString("amiiboSeries"));
-                                    gameModel.setCharacter(jsonObject1.getString("character"));
-                                    gameModel.setGameSeries(jsonObject1.getString("gameSeries"));
-                                    gameModel.setHead(jsonObject1.getString("head"));
-                                    gameModel.setImage(jsonObject1.getString("image"));
-                                    gameModel.setName(jsonObject1.getString("name"));
-                                    gameModel.setTail(jsonObject1.getString("tail"));
-                                    gameModel.setType(jsonObject1.getString("type"));*/
-
-                                    /*gameModel.setAu(jsonObjectRelease.getString("au"));
-                                    gameModel.setEu(jsonObjectRelease.getString("eu"));
-                                    gameModel.setJp(jsonObjectRelease.getString("jp"));
-                                    gameModel.setNa(jsonObjectRelease.getString("na"));*/
-                                    Log.e("Release au", jsonObjectRelease.getString("au"));
-                                    //  Log.e("Relase", jsonObject1.getString("release"));
-                                    arrayListGame.add(gameModel);
-
-                                } catch (JSONException e) {
-                                    e.printStackTrace();
-                                }
-
-
-                            }
-                            runOnUiThread(new Runnable() {
-                                @Override
-                                public void run() {
-                                    llNoData.setVisibility(View.GONE);
-                                    recyclerView.setVisibility(View.VISIBLE);
-                                    recyclerView.setAdapter(new GameAdapter(HomeRetrofitActivity.this, arrayListGame));
-                                }
-                            });
-                        } else {
-                            llNoData.setVisibility(View.VISIBLE);
-                            recyclerView.setVisibility(View.GONE);
-                        }
-                    }
-                });
-            }
-        } catch (JSONException e) {
-            e.printStackTrace();
-            llNoData.setVisibility(View.VISIBLE);
-            recyclerView.setVisibility(View.GONE);
-        }
     }
 
     @Override
