+++ app/src/main/java/in/khatri/rahul/amiiboapp/java/retrofit/utils/ApiInterface.java
@@ -1,10 +1,15 @@
 package in.khatri.rahul.amiiboapp.java.retrofit.utils;
 
 import com.google.gson.JsonElement;
+
+import java.util.ArrayList;
+
+import in.khatri.rahul.amiiboapp.java.retrofit.model.GameDataList;
+import in.khatri.rahul.amiiboapp.java.retrofit.model.GameRetrofitModel;
 import retrofit2.Call;
 import retrofit2.http.GET;
 
 public interface ApiInterface {
     @GET("amiibo")
-    Call<JsonElement> getGameData();
+    Call<GameDataList> getGameData();
 }
